+++ hass-actronque/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [v2026.2.1] - 2026-02-02
 
 ### Added
-- 
+- _jsonSettings static field with optimized settings
+- string pooling.
 
 ### Removed
 -